Split off some classes from molecule.py into components.py
